video: tegra: dc: Fix NULL address access
tegra_dc_get_numof_dispheads can return negative
value for errors. Since idx is unsigned, automatic
type promotion can cause the negative value to be
intepreted as large postive value. This caused if
condition to be true and causing invalid
memory access to occur in tegra_dcs

Bug 2331746

Reviewed-on: https://git-master.nvidia.com/r/1852200
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/video/tegra/dc/dc.c
@@ -3450,7 +3450,7 @@ EXPORT_SYMBOL(tegra_dc_get_stride);
 
 struct tegra_dc *tegra_dc_get_dc(unsigned idx)
 {
-	if (idx < tegra_dc_get_numof_dispheads())
+	if ((int)idx < tegra_dc_get_numof_dispheads())
 		return tegra_dcs[idx];
 	else
 		return NULL;
